Refactor Dockerfile: install npm dependencies on image build time and only run youtube-to-anchorfm when container is starting
This is a step towards making the docker image usable for one-off running

Note: the part where dependencies are being installed with apt-get is refactored to remove some duplicate and to merge some RUN opeartions in to single RUN operation and to clean the cache of apt-get. Cleaning of the cache was recommended by sonar. Also, --reinstall was removed, not sure what was the purpose of adding it.