Implement logging to file and create a png screenshot file in the directory where logs are kept
The logs are configured to be kept in `./logs`. The screenshot is configured to be saved in `./logs/screenshot.png`. This change will be useful in the scenario where youtube-to-anchorfm is run with the docker image. IT should be possible to mount a volume in order to see the logs and the screenshot. Note that the log format is the same in the console, i.e the users won't notice a difference when using youtube-to-anchorfm with github action.